Refactor Layout.
